+++ Tests/IntegrationTest.php
@@ -12,10 +12,9 @@
 namespace Symfony\AI\Store\Bridge\Neo4j\Tests;
 
 use PHPUnit\Framework\Attributes\Group;
-use Symfony\AI\Store\Bridge\Neo4j\Store;
+use Symfony\AI\Store\Bridge\Neo4j\StoreFactory;
 use Symfony\AI\Store\StoreInterface;
 use Symfony\AI\Store\Test\AbstractStoreIntegrationTestCase;
-use Symfony\Component\HttpClient\HttpClient;
 
 /**
  * @author Christopher Hertel <mail@christopher-hertel.de>
@@ -34,14 +33,13 @@ final class IntegrationTest extends AbstractStoreIntegrationTestCase
 
     protected static function createStore(): StoreInterface
     {
-        return new Store(
-            HttpClient::create(),
-            'http://127.0.0.1:7474',
-            'neo4j',
-            'symfonyai',
+        return StoreFactory::create(
             'neo4j',
             'test_vector_index',
             'TestNode',
+            'http://127.0.0.1:7474',
+            'neo4j',
+            'symfonyai',
             embeddingsDimension: 3,
         );
     }